+++ python/GafferImageUI/ImageReaderUI.py
@@ -44,6 +44,8 @@ from GafferUI.PlugValueWidget import sole
 
 from . import OpenColorIOTransformUI
 
+from GafferUI.PlugValueWidget import sole
+
 Gaffer.Metadata.registerNode(
 
 	GafferImage.ImageReader,
@@ -255,6 +257,19 @@ Gaffer.Metadata.registerNode(
 
 		],
 
+		"availableFrames" : [
+
+			"description",
+			"""
+			A list of the available frames for the current file sequence.
+			Empty when the input `fileName` is not a file sequence.
+			""",
+
+			"layout:section", "Frames",
+			"plugValueWidget:type", "GafferImageUI.ImageReaderUI._AvailableFramesPlugValueWidget",
+
+		],
+
 	}
 
 )
@@ -303,3 +318,20 @@ class _ColorSpacePlugValueWidget( GafferUI.PresetsPlugValueWidget ) :
 		node = plug.node()
 		if isinstance( node, GafferImage.ImageReader ) :
 			return [ node["__intermediateColorSpace"] ]
+
+class _AvailableFramesPlugValueWidget( GafferUI.PlugValueWidget ) :
+
+	def __init__( self, plug, **kw ) :
+
+		self.__textWidget = GafferUI.TextWidget( editable = False )
+		GafferUI.PlugValueWidget.__init__( self, self.__textWidget, plug, **kw )
+
+	def _updateFromValues( self, values, exception ) :
+
+		value = sole( values )
+		if value is None :
+			self.__textWidget.setText( "---" )
+		else :
+			self.__textWidget.setText( str( IECore.frameListFromList( list( value ) ) ) )
+
+		self.__textWidget.setErrored( exception is not None )